app/xinit: make startx's $? a useful value
When startx fails to start X, it's most likely xinit that failed.  xinit
returns a proper return code (1), but that gets clobbed in the startx
script by clean-up commands.  This patch saves $? from xinit and forces
startx to exit with that value.

This way, if startx actually fails to start X, $? reflects that.

Signed-off-by: Andres Salomon <dilinger@debian.org>