Simplify environment juggling by using fork() instead of vfork()
xinit used to copy process environment in order to give client correct DISPLAY
and WINDOWPATH variables. It was not possible to do it in client process because
it was vforked, not forked.

As vfork() usage was not entirely correct (manpage specifies that behaviour is
undefined if there are any memory modifications between vfork and exec), switch
to fork(), move setting environment variables to client process and and drop all
the manual mucking with environment.

Signed-off-by: Mikhail Gusarov <dottedmag@dottedmag.net>
Signed-off-by: Alan Coopersmith <alan.coopersmith@oracle.com>