Use default clientargs, serverargs, and display if none were set
Previously, we did not use the default values unless server or client weren't
set, but we should still use the defaults if they were not set but the server
was.  This is most evident when you want to tell startx which server to use,
but you want startx to figure out which display to use automatically.

This fixes a regression introduced by the previous patch on XQuartz:
http://xquartz.macosforge.org/trac/ticket/523

Signed-off-by: Jeremy Huddleston <jeremyhu@apple.com>